Crate layout_api

This module contains traits in script used generically in the rest of Servo. The traits are here instead of in script so that these modules won’t have to depend on script.

Modules

layout_damage 🔒

layout_dom 🔒

layout_element 🔒

layout_node 🔒

pseudo_element_chain 🔒

Structs

AccessibilityDamage

AnimatingImages

AxesOverflow

DomParallelInfo

Information that we need stored in each DOM node.

HTMLCanvasData

HTMLMediaData

HitTestFlags

HitTestResult

The result of a hit test query.

HitTestResultItem

Describe an item that matched a hit-test query.

IFrameSize

ImageAnimationState

LayoutConfig

LayoutDamage

Individual layout actions that may be necessary after restyling. This is an extension of RestyleDamage from stylo, which only uses the 4 lower bits.

MediaFrame

MediaMetadata

OffsetParentResponse

PendingImage

The data associated with an image that is not yet present in the image cache. Used by the script thread to hold on to DOM elements that need to be repainted when an image fetch is complete.

PendingRasterizationImage

A data structure to track vector image that are fully loaded (i.e has a parsed SVG tree) but not yet rasterized to the size needed by layout. The rasterization is happening in the image cache.

PendingRestyle

A pending restyle.

PhysicalSides

PseudoElementChain

A chain of pseudo-elements up to two levels deep. This is used to represent cases where a pseudo-element has its own child pseudo element (for instance .div::after::marker). If both Self::primary and Self::secondary are None, then this chain represents the element itself. Not all combinations of pseudo-elements are possible and we may not be able to calculate a style for all PseudoElementChains.

ReflowPhasesRun

The phases of reflow that were run when processing a reflow in layout.

ReflowRequest

Information needed for a script-initiated reflow.

ReflowRequestRestyle

Information needed for a script-initiated reflow that requires a restyle and reconstruction of box and fragment trees.

ReflowResult

Information derived from a layout pass that needs to be returned to the script thread.

ReflowStatistics

RestyleReason

Conditions which cause a [Document] to need to be restyled during reflow, which might cause the rest of layout to happen as well.

SVGElementData

ScrollContainerQueryFlags

StyleData

ThreadStateRestorer 🔒

TrustedNodeAddress

The address of a node known to be valid. These are sent from script to layout.

Enums

BoxAreaType

Type of the area of CSS box for query. See https://www.w3.org/TR/css-box-3/#box-model.

FragmentType

The type of fragment that a scroll root is created for.

LayoutElementType

LayoutImageDestination

The destination in layout where an image is needed.

LayoutNodeType

NodeRenderingType

Whether or not this node is being rendered or delegates rendering according to the HTML standard.

PendingImageState

Whether the pending image needs to be fetched or is waiting on an existing fetch.

QueryMsg

ReflowGoal

The goal of a reflow request.

ScrollContainerResponse

Traits

DangerousStyleElement

An element that can be passed to stylo and selectors that allows accessing the parent node. We consider this to be too dangerous for normal layout, so it is reserved only for using stylo and selectors.

DangerousStyleNode

A node that can be passed to stylo and selectors that allows accessing the parent node. We consider this to be too dangerous for normal layout, so it is reserved only for using stylo and selectors.

GenericLayoutDataTrait

Layout

LayoutDataTrait

LayoutDomTypeBundle

A trait that holds all the concrete implementations of the Layout DOM traits. This is useful because it means that other types (specifically the implementation of the Layout trait) can be parameterized over a single type (LayoutDomTypeBundle) rather than all of the various Layout DOM trait implementations.

LayoutElement

A trait that exposes a DOM element to layout. Implementors of this trait must abide by certain safety requirements. Layout will only ever access and mutate each element from a single thread at a time, though children may be used in parallel from other threads. That is why this trait does not allow access to parent nodes, as it would make it easy to cause race conditions and memory errors.

LayoutFactory

LayoutNode

A trait that exposes a DOM nodes to layout. Implementors of this trait must abide by certain safety requirements. Layout will only ever access and mutate each node from a single thread at a time, though children may be used in parallel from other threads. That is why this trait does not allow access to parent nodes, as it would make it easy to cause race conditions and memory errors.

ScriptThreadFactory

This trait is part of layout_api because it depends on both script_traits and also LayoutFactory from this crate. If it was in script_traits there would be a circular dependency.

Functions

combine_id_with_fragment_type

node_id_from_scroll_id

with_layout_state

Set up the thread-local state to reflect that layout code is about to run, then call the provided function. This must be used when running code that will interact with the DOM tree through types like ServoLayoutNode, ServoLayoutElement, and LayoutDom, which have rules about how they must be used from layout worker threads.

Type Aliases

CSSPixelRectVec

DangerousStyleElementOf

Type alias to extract ConcreteDangerousStyleElement from a LayoutDomTypeBundle implementation.

DangerousStyleNodeOf

Type alias to extract ConcreteDangerousStyleNode from a LayoutDomTypeBundle implementation.

GenericLayoutData

IFrameSizes

LayoutElementOf

Type alias to extract ConcreteLayoutElement from a LayoutDomTypeBundle implementation.

LayoutNodeOf

Type alias to extract ConcreteLayoutNode from a LayoutDomTypeBundle implementation.