layout_api::layout_node

Trait DangerousStyleNode

pub trait DangerousStyleNode<'dom>:
    TNode
    + Sized
    + NodeInfo
    + Send
    + Sync {
    type ConcreteTypeBundle: LayoutDomTypeBundle<'dom>;

    // Required method
    fn layout_node(&self) -> LayoutNodeOf<'dom, Self::ConcreteTypeBundle>;
}

A node that can be passed to stylo and selectors that allows accessing the parent node. We consider this to be too dangerous for normal layout, so it is reserved only for using stylo and selectors.

If you are not interfacing with stylo and selectors you should not use this type, unless you know what you are doing.

Required Associated Types

type ConcreteTypeBundle: LayoutDomTypeBundle<'dom>

The concrete implementation of LayoutDomTypeBundle implemented in script.

Required Methods

fn layout_node(&self) -> LayoutNodeOf<'dom, Self::ConcreteTypeBundle>

Get a handle to the original “safe” version of this node, a LayoutNode implementation.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.