layout_api

Trait LayoutDomTypeBundle

pub trait LayoutDomTypeBundle<'dom> {
    type ConcreteLayoutNode: LayoutNode<'dom>;
    type ConcreteLayoutElement: LayoutElement<'dom>;
    type ConcreteDangerousStyleNode: DangerousStyleNode<'dom>;
    type ConcreteDangerousStyleElement: DangerousStyleElement<'dom>;
}

A trait that holds all the concrete implementations of the Layout DOM traits. This is useful because it means that other types (specifically the implementation of the Layout trait) can be parameterized over a single type (LayoutDomTypeBundle) rather than all of the various Layout DOM trait implementations.

Required Associated Types

type ConcreteLayoutNode: LayoutNode<'dom>

The concrete implementation of LayoutNode from script.

type ConcreteLayoutElement: LayoutElement<'dom>

The concrete implementation of LayoutElement from script.

type ConcreteDangerousStyleNode: DangerousStyleNode<'dom>

The concrete implementation of DangerousStyleNode from script.

type ConcreteDangerousStyleElement: DangerousStyleElement<'dom>

The concrete implementation of DangerousStyleElement from script.