layout_api::wrapper_traits

Trait ThreadSafeLayoutElement

pub trait ThreadSafeLayoutElement<'dom>:
    Clone
    + Copy
    + Sized
    + Debug
    + Element<Impl = SelectorImpl> {
    type ConcreteThreadSafeLayoutNode: ThreadSafeLayoutNode<'dom, ConcreteThreadSafeLayoutElement = Self>;
    type ConcreteElement: TElement;

    // Required methods
    fn as_node(&self) -> Self::ConcreteThreadSafeLayoutNode;
    fn with_pseudo(&self, pseudo: PseudoElement) -> Option<Self>;
    fn type_id(&self) -> Option<LayoutNodeType>;
    fn unsafe_get(self) -> Self::ConcreteElement;
    fn get_local_name(&self) -> &LocalName;
    fn get_attr(&self, namespace: &Namespace, name: &LocalName) -> Option<&str>;
    fn get_attr_enum(
        &self,
        namespace: &Namespace,
        name: &LocalName,
    ) -> Option<&AttrValue>;
    fn style_data(&self) -> AtomicRef<'_, ElementData>;
    fn pseudo_element_chain(&self) -> PseudoElementChain;
    fn is_shadow_host(&self) -> bool;
    fn is_body_element_of_html_element_root(&self) -> bool;
    fn is_root(&self) -> bool;

    // Provided method
    fn style(&self, context: &SharedStyleContext<'_>) -> Arc<ComputedValues> { ... }

}

Required Associated Types

type ConcreteThreadSafeLayoutNode: ThreadSafeLayoutNode<'dom, ConcreteThreadSafeLayoutElement = Self>

type ConcreteElement: TElement

This type alias is just a work-around to avoid writing

<Self::ConcreteThreadSafeLayoutNode as ThreadSafeLayoutNode>::ConcreteElement

Required Methods

fn as_node(&self) -> Self::ConcreteThreadSafeLayoutNode

fn with_pseudo(&self, pseudo: PseudoElement) -> Option<Self>

Creates a new ThreadSafeLayoutElement for the same LayoutElement with a different pseudo-element type.

Returns None if this pseudo doesn’t apply to the given element for one of the following reasons:

pseudo is eager and is not defined in the stylesheet. In this case, there is not reason to process the pseudo element at all.

pseudo is for ::servo-details-summary or ::servo-details-content and it doesn’t apply to this element, either because it isn’t a details or is in the wrong state.

fn type_id(&self) -> Option<LayoutNodeType>

Returns the type ID of this node. Returns None if this is a pseudo-element; otherwise, returns Some.

fn unsafe_get(self) -> Self::ConcreteElement

Returns access to the underlying TElement. This is breaks the abstraction barrier of ThreadSafeLayout wrapper layer, and can lead to races if not used carefully.

We need this so that the functions defined on this trait can call lazily_compute_pseudo_element_style, which operates on TElement.

fn get_local_name(&self) -> &LocalName

Get the local name of this element. See https://dom.spec.whatwg.org/#concept-element-local-name.

fn get_attr(&self, namespace: &Namespace, name: &LocalName) -> Option<&str>

fn get_attr_enum( &self, namespace: &Namespace, name: &LocalName, ) -> Option<&AttrValue>

fn style_data(&self) -> AtomicRef<'_, ElementData>

fn pseudo_element_chain(&self) -> PseudoElementChain

fn is_shadow_host(&self) -> bool

fn is_body_element_of_html_element_root(&self) -> bool

Returns whether this node is a body element of an html element root in an HTML element document.

Note that this does require accessing the parent, which this interface technically forbids. But accessing the parent is only unsafe insofar as it can be used to reach siblings and cousins. A simple immutable borrow of the parent data is fine, since the bottom-up traversal will not process the parent until all the children have been processed.

fn is_root(&self) -> bool

Returns whether this node is the root element in an HTML document element.

Note that, like Self::is_body_element_of_html_element_root, this accesses the parent. As in that case, since this is an immutable borrow, we do not violate thread safety.

Provided Methods

fn style(&self, context: &SharedStyleContext<'_>) -> Arc<ComputedValues>

Returns the style results for the given node. If CSS selector matching has not yet been performed, fails.

Unlike the version on TNode, this handles pseudo-elements.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.