layout_api::wrapper_traits

Trait ThreadSafeLayoutNode

pub trait ThreadSafeLayoutNode<'dom>:
    Clone
    + Copy
    + Debug
    + NodeInfo
    + PartialEq
    + Sized {
    type ConcreteNode: LayoutNode<'dom, ConcreteThreadSafeLayoutNode = Self>;
    type ConcreteElement: TElement;
    type ConcreteThreadSafeLayoutElement: ThreadSafeLayoutElement<'dom, ConcreteThreadSafeLayoutNode = Self> + Element<Impl = SelectorImpl>;
    type ChildrenIterator: Iterator<Item = Self> + Sized;

    // Required methods
    fn opaque(&self) -> OpaqueNode;
    fn type_id(&self) -> Option<LayoutNodeType>;
    fn parent_style(&self) -> Arc<ComputedValues>;
    fn initialize_layout_data<RequestedLayoutDataType: LayoutDataTrait>(&self);
    fn debug_id(self) -> usize;
    fn children(&self) -> LayoutIterator<Self::ChildrenIterator>;
    fn as_element(&self) -> Option<Self::ConcreteThreadSafeLayoutElement>;
    fn as_html_element(&self) -> Option<Self::ConcreteThreadSafeLayoutElement>;
    fn style_data(&self) -> Option<&'dom StyleData>;
    fn layout_data(&self) -> Option<&'dom GenericLayoutData>;
    fn unsafe_get(self) -> Self::ConcreteNode;
    fn text_content(self) -> Cow<'dom, str>;
    fn selection(&self) -> Option<TextByteRange>;
    fn image_url(&self) -> Option<ServoUrl>;
    fn image_density(&self) -> Option<f64>;
    fn image_data(&self) -> Option<(Option<Image>, Option<ImageMetadata>)>;
    fn showing_broken_image_icon(&self) -> bool;
    fn canvas_data(&self) -> Option<HTMLCanvasData>;
    fn svg_data(&self) -> Option<SVGElementData>;
    fn media_data(&self) -> Option<HTMLMediaData>;
    fn iframe_browsing_context_id(&self) -> Option<BrowsingContextId>;
    fn iframe_pipeline_id(&self) -> Option<PipelineId>;
    fn get_span(&self) -> Option<u32>;
    fn get_colspan(&self) -> Option<u32>;
    fn get_rowspan(&self) -> Option<u32>;
    fn pseudo_element_chain(&self) -> PseudoElementChain;
    fn with_pseudo_element_chain(
        &self,
        pseudo_element_chain: PseudoElementChain,
    ) -> Self;
    fn set_uses_content_attribute_with_attr(
        &self,
        _uses_content_attribute_with_attr: bool,
    );

    // Provided methods
    fn style(&self, context: &SharedStyleContext<'_>) -> Arc<ComputedValues> { ... }
    fn is_content(&self) -> bool { ... }
    fn with_pseudo(&self, pseudo_element_type: PseudoElement) -> Option<Self> { ... }

}

A thread-safe version of LayoutNode, used during flow construction. This type of layout node does not allow any parents or siblings of nodes to be accessed, to avoid races.

Required Associated Types

type ConcreteNode: LayoutNode<'dom, ConcreteThreadSafeLayoutNode = Self>

type ConcreteElement: TElement

type ConcreteThreadSafeLayoutElement: ThreadSafeLayoutElement<'dom, ConcreteThreadSafeLayoutNode = Self> + Element<Impl = SelectorImpl>

type ChildrenIterator: Iterator<Item = Self> + Sized

Required Methods

fn opaque(&self) -> OpaqueNode

Converts self into an OpaqueNode.

fn type_id(&self) -> Option<LayoutNodeType>

Returns the type ID of this node. Returns None if this is a pseudo-element; otherwise, returns Some.

fn parent_style(&self) -> Arc<ComputedValues>

Returns the style for a text node. This is computed on the fly from the parent style to avoid traversing text nodes in the style system.

Note that this does require accessing the parent, which this interface technically forbids. But accessing the parent is only unsafe insofar as it can be used to reach siblings and cousins. A simple immutable borrow of the parent data is fine, since the bottom-up traversal will not process the parent until all the children have been processed.

fn initialize_layout_data<RequestedLayoutDataType: LayoutDataTrait>(&self)

Initialize this node with empty opaque layout data.

Safety

This method is unsafe because it modifies the given node during layout. Callers should ensure that no other layout thread is attempting to read or modify the opaque layout data of this node.

fn debug_id(self) -> usize

fn children(&self) -> LayoutIterator<Self::ChildrenIterator>

Returns an iterator over this node’s children.

fn as_element(&self) -> Option<Self::ConcreteThreadSafeLayoutElement>

Returns a ThreadSafeLayoutElement if this is an element, None otherwise.

fn as_html_element(&self) -> Option<Self::ConcreteThreadSafeLayoutElement>

Returns a ThreadSafeLayoutElement if this is an element in an HTML namespace, None otherwise.

fn style_data(&self) -> Option<&'dom StyleData>

Get the StyleData for this node. Returns None if the node is unstyled.

fn layout_data(&self) -> Option<&'dom GenericLayoutData>

Get the layout data of this node, attempting to downcast it to the desired type. Returns None if there is no layout data or it isn’t of the desired type.

fn unsafe_get(self) -> Self::ConcreteNode

Returns access to the underlying LayoutNode. This is breaks the abstraction barrier of ThreadSafeLayout wrapper layer, and can lead to races if not used carefully.

We need this because the implementation of some methods need to access the layout data flags, and we have this annoying trait separation between script and layout :-(

fn text_content(self) -> Cow<'dom, str>

Get the text content of this node, if it is a text node.

Note: This should only be called on text nodes.

fn selection(&self) -> Option<TextByteRange>

If selection intersects this node, return it. Otherwise, returns None.

fn image_url(&self) -> Option<ServoUrl>

If this is an image element, returns its URL. If this is not an image element, fails.

fn image_density(&self) -> Option<f64>

If this is an image element, returns its current-pixel-density. If this is not an image element, fails.

fn image_data(&self) -> Option<(Option<Image>, Option<ImageMetadata>)>

If this is an image element, returns its image data. Otherwise, returns None.

fn showing_broken_image_icon(&self) -> bool

Whether or not this is an image element that is showing a broken image icon.

fn canvas_data(&self) -> Option<HTMLCanvasData>

fn svg_data(&self) -> Option<SVGElementData>

fn media_data(&self) -> Option<HTMLMediaData>

fn iframe_browsing_context_id(&self) -> Option<BrowsingContextId>

If this node is an iframe element, returns its browsing context ID. If this node is not an iframe element, fails. Returns None if there is no nested browsing context.

fn iframe_pipeline_id(&self) -> Option<PipelineId>

If this node is an iframe element, returns its pipeline ID. If this node is not an iframe element, fails. Returns None if there is no nested browsing context.

fn get_span(&self) -> Option<u32>

fn get_colspan(&self) -> Option<u32>

fn get_rowspan(&self) -> Option<u32>

fn pseudo_element_chain(&self) -> PseudoElementChain

fn with_pseudo_element_chain( &self, pseudo_element_chain: PseudoElementChain, ) -> Self

fn set_uses_content_attribute_with_attr( &self, _uses_content_attribute_with_attr: bool, )

Set whether or not this node has an active pseudo-element style with a content attribute that uses attr.

Safety

This function accesses and modifies the underlying DOM object and should not be used by more than a single thread at once.

Provided Methods

fn style(&self, context: &SharedStyleContext<'_>) -> Arc<ComputedValues>

fn is_content(&self) -> bool

Returns true if this node contributes content. This is used in the implementation of empty_cells per CSS 2.1  17.6.1.1.

fn with_pseudo(&self, pseudo_element_type: PseudoElement) -> Option<Self>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.